replace accidental floats with double

+++ src/lib/noise.cc
@@ -105,10 +105,10 @@ static constexpr double HEURISTIC_3D = 32.696;
 static constexpr double F3 = 1.0 / 3.0;
 static constexpr double G3 = 1.0 / 6.0;
 
-static float grad3(int32_t hash, float x, float y, float z) {
+static double grad3(int32_t hash, double x, double y, double z) {
     int h = hash & 15;     // Convert low 4 bits of hash code into 12 simple
-    float u = h < 8 ? x : y; // gradient directions, and compute dot product.
-    float v = h < 4 ? y : h == 12 || h == 14 ? x : z; // Fix repeats at h = 12 to 15
+    double u = h < 8 ? x : y; // gradient directions, and compute dot product.
+    double v = h < 4 ? y : h == 12 || h == 14 ? x : z; // Fix repeats at h = 12 to 15
     return ((h & 1) ? -u : u) + ((h & 2) ? -v : v);
 }
